+++ src/qtouch/qtouch.h
@@ -29,11 +29,6 @@ extern "C" {
  *----------------------------------------------------------------------------*/
 
 #include "touch_api_ptc.h"
-#include <platform_config.h>
-
-#if PLATFORM_BITBOXBASE == 1
-bool qtouch_get_button_state(size_t idx);
-#endif
 
 /**********************************************************/
 /******************* Acquisition controls *****************/
@@ -75,206 +70,6 @@ bool qtouch_get_button_state(size_t idx);
 /*----------------------------------------------------------------------------
  *     defines
  *----------------------------------------------------------------------------*/
-
-#if PLATFORM_BITBOX02 == 1
-/**********************************************************/
-/***************** Node Params   ******************/
-/**********************************************************/
-/* Acquisition Set 1 */
-/* Defines the number of sensor nodes in the acquisition set
- * Range: 1 to 65535.
- * Default value: 1
- */
-#define DEF_NUM_CHANNELS (8)
-
-/* Defines node parameter setting
- * {X-line, Y-line, Charge Share Delay, NODE_RSEL_PRSC(series resistor, prescaler), NODE_G(Analog
- * Gain , Digital Gain), filter level}
- */
-// Slider 1 buttons
-#define NODE_0_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(26), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_1), \
-            NODE_GAIN(GAIN_4, GAIN_4), FILTER_LEVEL_512                     \
-    }
-#define NODE_1_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(27), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_1), \
-            NODE_GAIN(GAIN_4, GAIN_4), FILTER_LEVEL_512                     \
-    }
-#define NODE_2_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(28), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_1), \
-            NODE_GAIN(GAIN_4, GAIN_4), FILTER_LEVEL_512                     \
-    }
-#define NODE_3_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(29), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_1), \
-            NODE_GAIN(GAIN_4, GAIN_4), FILTER_LEVEL_512                     \
-    }
-// Slider 0 buttons
-#define NODE_4_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(30), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_1), \
-            NODE_GAIN(GAIN_4, GAIN_4), FILTER_LEVEL_512                     \
-    }
-#define NODE_5_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(31), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_1), \
-            NODE_GAIN(GAIN_4, GAIN_4), FILTER_LEVEL_512                     \
-    }
-#define NODE_6_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(20), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_1), \
-            NODE_GAIN(GAIN_4, GAIN_4), FILTER_LEVEL_512                     \
-    }
-#define NODE_7_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(21), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_1), \
-            NODE_GAIN(GAIN_4, GAIN_4), FILTER_LEVEL_512                     \
-    }
-
-/**********************************************************/
-/***************** Key Params   ******************/
-/**********************************************************/
-/* Defines the number of key sensors
- * Range: 1 to 65535.
- * Default value: 1
- */
-#define DEF_NUM_SENSORS (DEF_NUM_CHANNELS)
-
-/* Defines Key Sensor setting
- * {Sensor Threshold, Sensor Hysterisis, Sensor AKS}
- */
-// 0..3 higher Slider left to right 4..7 lower Slider right to left
-#define KEY_0_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_1_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_2_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_3_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_4_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_5_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_6_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_7_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-
-#elif PLATFORM_BITBOXBASE == 1
-/**********************************************************/
-/***************** Node Params   ******************/
-/**********************************************************/
-/* Acquisition Set 1 */
-/* Defines the number of sensor nodes in the acquisition set
- * Range: 1 to 65535.
- * Default value: 1
- */
-#define DEF_NUM_CHANNELS (6)
-
-/* Defines node parameter setting
- * {X-line, Y-line, Charge Share Delay, NODE_RSEL_PRSC(series resistor, prescaler), NODE_G(Analog
- * Gain , Digital Gain), filter level}
- */
-// Slider buttons
-#define NODE_0_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(23), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_2), \
-            NODE_GAIN(GAIN_8, GAIN_1), FILTER_LEVEL_32                      \
-    }
-#define NODE_1_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(22), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_2), \
-            NODE_GAIN(GAIN_8, GAIN_1), FILTER_LEVEL_32                      \
-    }
-#define NODE_2_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(21), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_2), \
-            NODE_GAIN(GAIN_8, GAIN_1), FILTER_LEVEL_32                      \
-    }
-#define NODE_3_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(20), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_2), \
-            NODE_GAIN(GAIN_8, GAIN_1), FILTER_LEVEL_32                      \
-    }
-// Top buttons
-#define NODE_4_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(30), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_4), \
-            NODE_GAIN(GAIN_16, GAIN_2), FILTER_LEVEL_256                    \
-    }
-#define NODE_5_PARAMS                                                       \
-    {                                                                       \
-        X_NONE, Y_LINE(31), 0, NODE_RSEL_PRSC(RSEL_VAL_20, PRSC_DIV_SEL_4), \
-            NODE_GAIN(GAIN_16, GAIN_2), FILTER_LEVEL_256                    \
-    }
-
-/**********************************************************/
-/***************** Key Params   ******************/
-/**********************************************************/
-/* Defines the number of key sensors
- * Range: 1 to 65535.
- * Default value: 1
- */
-#define DEF_NUM_SENSORS (DEF_NUM_CHANNELS)
-
-/* Defines Key Sensor setting
- * {Sensor Threshold, Sensor Hysterisis, Sensor AKS}
- */
-// 0..3 higher Slider left to right 4..7 lower Slider right to left
-#define KEY_0_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_1_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_2_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_3_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_4_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-#define KEY_5_PARAMS              \
-    {                             \
-        16, HYST_50, NO_AKS_GROUP \
-    }
-
-/*
- * Defines number of buttson
- */
-#define DEF_NUM_BUTTONS 2
-/*
- * Defines the offset to the first button in the node list
- */
-#define DEF_BUTTON_OFFSET 4
-#endif
 
 /* De-bounce counter for additional measurements to confirm touch detection
  * Range: 0 to 255.
@@ -347,17 +142,6 @@ bool qtouch_get_button_state(size_t idx);
  * This allows low noise button readings while keeping
  * fast responsiveness.
  */
-#if PLATFORM_BITBOX02 == 1
-#define DEF_NUM_SCROLLERS 2 // Number of scrollers (sliders or wheels)
-#define DEF_SCROLLER_NUM_CHANNELS 4 // Number of channels per scroller
-#define DEF_SCROLLER_OFFSET_0 4 // Index of first button in scroller
-#define DEF_SCROLLER_OFFSET_1 0 // Index of first button in scroller
-#elif PLATFORM_BITBOXBASE == 1
-#define DEF_NUM_SCROLLERS 1 // Number of scrollers (sliders or wheels)
-#define DEF_SCROLLER_NUM_CHANNELS 4 // Number of channels per scroller
-#define DEF_SCROLLER_OFFSET_0 0 // Index of first button in scroller
-#define DEF_SCROLLER_OFFSET_1 0 // Index of first button in scroller
-#endif
 #define DEF_SCROLLER_RESOLUTION 256 // Scroller resolution in bits
 #define DEF_SCROLLER_DET_THRESHOLD 25 // Scroller detect threshold
 #define DEF_SCROLLER_TOUCH_THRESHOLD 25 // Scroller active threshold
