+++ src/touch/gestures_bitboxbase.h
@@ -1,0 +1,25 @@
+// Copyright 2019 Shift Cryptosecurity AG
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef __GESTURES_BITBOXBASE_H
+#define __GESTURES_BITBOXBASE_H
+
+enum bitboxbase_button_id_t {
+    BITBOXBASE_BUTTON_LEFT,
+    BITBOXBASE_BUTTON_RIGHT,
+};
+
+enum bitboxbase_button_id_t gestures_button_which(const event_t* event);
+
+#endif // __GESTURES_BITBOXBASE_H
