+++ src/ui/components/confirm_gesture.c
@@ -18,7 +18,6 @@
 #include "ui_images.h"
 
 #include <hardfault.h>
-#include <platform/platform_config.h>
 #include <screen.h>
 #include <touch/gestures.h>
 #include <ui/ui_util.h>
