+++ src/ui/components/ui_images.c
@@ -14,7 +14,6 @@
 
 #include "ui_images.h"
 
-#include <platform/platform_config.h>
 #include <ui/ugui/ugui.h>
 
 void image_arrow(int x, int y, int height, arrow_orientation_t orientation)
