+++ test/unit-test/CMakeLists.txt
@@ -20,6 +20,7 @@ set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wno-unused-parameter -Wno-missing-prototype
 
 set(DBB-FILTERED-SOURCES
   ${DBB-FIRMWARE-COMMON-SOURCES}
+  ${DBB-FIRMWARE-UI-COMMON-SOURCES}
   ${DBB-FIRMWARE-UI-SOURCES}
   ${DBB-FIRMWARE-PROTOBUF-SOURCES}
   ${FIRMWARE-U2F-SOURCES}
@@ -46,6 +47,8 @@ find_package(Nanopb REQUIRED)
 
 add_library(bitbox
   STATIC
+  ${DBB-FIRMWARE-COMPONENT-SOURCES}
+  ${DBB-FIRMWARE-WORKFLOW-SOURCES}
   ${APP-BTC-SOURCES}
   ${DBB-FILTERED-SOURCES}
   ${CTAES-SOURCES}
